Weather Icon 지정하는 부분 UI layer 로 변경
