+++ app/src/main/java/com/jin/jjinweather/layer/domain/model/weather/DailyWeather.kt
@@ -2,7 +2,7 @@ package com.jin.jjinweather.layer.domain.model.weather
 
 data class DailyWeather(
     val forecastDay: Long,
-    val iconResId: Int,
+    val iconCode: String,
     val minTemperature: Number,
     val maxTemperature: Number
 )
