+++ app/src/main/java/com/jin/jjinweather/layer/domain/model/weather/HourlyWeather.kt
@@ -2,6 +2,6 @@ package com.jin.jjinweather.layer.domain.model.weather
 
 data class HourlyWeather(
     val forecastTime: Long,
-    val iconResId: Int,
+    val iconCode: String,
     val temperature: Number
 )
