+++ app/src/main/java/com/jin/jjinweather/layer/ui/onboarding/WeatherContentUI.kt
@@ -20,6 +20,7 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
+import com.jin.jjinweather.R
 import com.jin.jjinweather.layer.domain.model.UiState
 import com.jin.jjinweather.layer.domain.model.weather.Weather
 
@@ -52,6 +53,8 @@ fun WeatherLoadingContent() {
 fun WeatherLoadedContent(weather: Weather, onNavigateToOnboarding: () -> Unit) {
     // success 한 후, Main 화면으로 자동으로 이동할 수 있도록 처리필요.
     // 임시 UI 작성. Main 화면으로 이동 되면 다듬어질 예정.
+    val currentWeatherIconRes = mapWeatherIconToDrawable(weather.iconCode)
+
     Scaffold(modifier = Modifier.fillMaxSize()) { innerPadding ->
         Box(
             modifier = Modifier
@@ -62,7 +65,7 @@ fun WeatherLoadedContent(weather: Weather, onNavigateToOnboarding: () -> Unit) {
                 Text("위치 : ${weather.cityName}")
                 Text("현재 온도 : ${weather.currentTemperature}")
                 Image(
-                    painter = painterResource(id = weather.iconResId),
+                    painter = painterResource(id = currentWeatherIconRes),
                     contentDescription = "current temp icon",
                     modifier = Modifier.size(64.dp)
                 )
@@ -73,9 +76,10 @@ fun WeatherLoadedContent(weather: Weather, onNavigateToOnboarding: () -> Unit) {
                     contentPadding = PaddingValues(horizontal = 16.dp)
                 ) {
                     items(weather.hourlyWeatherList.size) { index ->
+                        val hourlyWeatherIconRes = mapWeatherIconToDrawable(weather.hourlyWeatherList[index].iconCode)
                         Column {
                             Image(
-                                painter = painterResource(id = weather.hourlyWeatherList[index].iconResId),
+                                painter = painterResource(id = hourlyWeatherIconRes),
                                 contentDescription = "hourly temp icon",
                                 modifier = Modifier.size(32.dp)
                             )
@@ -90,9 +94,10 @@ fun WeatherLoadedContent(weather: Weather, onNavigateToOnboarding: () -> Unit) {
                     contentPadding = PaddingValues(horizontal = 16.dp)
                 ) {
                     items(weather.dailyWeatherList.size) { index ->
+                        val dailyWeatherIconRes = mapWeatherIconToDrawable(weather.dailyWeatherList[index].iconCode)
                         Column {
                             Image(
-                                painter = painterResource(id = weather.dailyWeatherList[index].iconResId),
+                                painter = painterResource(id = dailyWeatherIconRes),
                                 contentDescription = "hourly temp icon",
                                 modifier = Modifier.size(32.dp)
                             )
@@ -118,5 +123,22 @@ fun WeatherErrorContent(message: String) {
                 Text("Error : $message", color = Color.Red)
             }
         }
+    }
+}
+
+private fun mapWeatherIconToDrawable(icon: String): Int {
+    return when (icon) {
+        "01d" -> R.drawable.ic_main_clear_sky_day
+        "01n" -> R.drawable.ic_main_clear_sky_night
+        "02d" -> R.drawable.ic_main_few_clouds_day
+        "02n" -> R.drawable.ic_main_few_clouds_night
+        "03d", "03n", "04d", "04n" -> R.drawable.ic_main_scattered_clouds
+        "09d", "09n" -> R.drawable.ic_main_shower_rain
+        "10d" -> R.drawable.ic_main_rain_day
+        "10n" -> R.drawable.ic_main_rain_night
+        "11d", "11n" -> R.drawable.ic_main_thunderstorm
+        "13d", "13n" -> R.drawable.ic_main_snow
+        "50d", "50n" -> R.drawable.ic_main_mist
+        else -> R.drawable.ic_main_clear_sky_day
     }
 }
